Fix "_id" fields not being unique.
The field not being unique may cause issues with the difference checker, as it just randomly decides the order of certain objects, causing false positives in difference checkers.

- Make the "_id" fields use more than just the name, as duplicate names are possible in the cases of enemies.
- Add an "_id" field to transients and refactor the file a bit.

These are the only current files that have ambiguity, but there may be more cases of this issue possible later (think items, primarily). Might need to improve the "_id" field of those later, too. For now, however, this suffices.