Fix some "_id" fields not being unique. (#17)
* Fix "_id" fields not being unique.

The field not being unique may cause issues with the difference checker, as it just randomly decides the order of certain objects, causing false positives in difference checkers.

- Make the "_id" fields use more than just the name, as duplicate names are possible in the cases of enemies.
- Add an "_id" field to transients and refactor the file a bit.

These are the only current files that have ambiguity, but there may be more cases of this issue possible later (think items, primarily). Might need to improve the "_id" field of those later, too. For now, however, this suffices.

* Manual regeneration of the  2018-05-17 17:00 droptables.

* Update README to add the transient "_id" field.

* Concat strings and remove redundant null JSON properties.

* Manual regeneration of the 2018-05-17 17:00 droptables.

* Resolve duplicate _id fields still occurring.

Reverts the dynamic salt in 90% of the cases, so comparisons can be more easily made. Now, when the same ID has been found in an array, try to hash it with some properties until giving up eventually and just skipping adding the object when it's the _exact_ same. This applies to enemyModTables.js and modLocations.js, as those are the currently known cases of duplicates occurring.

Also refactors the other files having inconsistent formatting.

* Manual regeneration of the 2018-05-17 17:00 droptables.

* Revert "ememy" fix for now and decrease amount of _id changes.

This reverts my previous fix for ememy > enemy for now and discards all changes to the object already in the array when all previous checks fail. This causes less removals and additions when just de-duplicating the array as the original item doesn't get edited at all.

* Manual regeneration of the 2018-05-17 17:00 droptables.

* Add modular way to check hashes of arrays.

Adds checkUniqueHash(items, source, [...properties]) function. Expects "_id" field to exist within the source object, and experts the first property to be the value that's the default hash.

* Manual regeneration of the 2018-05-17 17:00 droptables.

* var -> let